requirements.txt
keras>=2.12.0
opencv-contrib-python-headless
tensorflow-cpu
streamlit_webrtc
SpeechRecognition==3.10.0
FuzzyTM>=0.4.0
blosc2~=2.0.0
cython>=0.29.21
llvmlite<0.41,>=0.40
tiktoken==0.3.1
gTTS==2.4.0
Janome==0.5.0
pykakasi
pygame